+++ elastic-job-lite/elastic-job-lite-console/src/main/java/com/dangdang/ddframe/job/lite/console/restful/RegistryCenterRestfulApi.java
@@ -15,58 +15,70 @@
  * </p>
  */
 
-package com.dangdang.ddframe.job.lite.console.controller;
+package com.dangdang.ddframe.job.lite.console.restful;
 
 import com.dangdang.ddframe.job.lite.console.domain.RegistryCenterConfiguration;
 import com.dangdang.ddframe.job.lite.console.service.RegistryCenterService;
+import com.dangdang.ddframe.job.lite.console.service.impl.RegistryCenterServiceImpl;
+import com.dangdang.ddframe.job.lite.console.util.SessionRegistryCenterConfiguration;
 import com.dangdang.ddframe.job.lite.lifecycle.internal.reg.RegistryCenterFactory;
 import com.dangdang.ddframe.job.reg.exception.RegException;
 import com.google.common.base.Optional;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
 
-import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
 import java.util.Collection;
 
-@RestController
-@RequestMapping("registry_center")
-public class RegistryCenterController {
+@Path("/registry_center")
+public class RegistryCenterRestfulApi {
     
     public static final String REG_CENTER_CONFIG_KEY = "reg_center_config_key";
     
-    @Resource
-    private RegistryCenterService regCenterService;
+    private RegistryCenterService regCenterService = new RegistryCenterServiceImpl();
     
-    @RequestMapping(method = RequestMethod.GET)
-    public Collection<RegistryCenterConfiguration> load(final HttpSession session) {
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    public Collection<RegistryCenterConfiguration> load(final @Context HttpServletRequest request) {
         Optional<RegistryCenterConfiguration> regCenterConfig = regCenterService.loadActivated();
         if (regCenterConfig.isPresent()) {
-            setRegistryCenterNameToSession(regCenterConfig.get(), session);
+            setRegistryCenterNameToSession(regCenterConfig.get(), request.getSession());
         }
         return regCenterService.loadAll();
     }
     
-    @RequestMapping(method = RequestMethod.POST)
+    @POST
+    @Produces(MediaType.APPLICATION_JSON)
     public boolean add(final RegistryCenterConfiguration config) {
         return regCenterService.add(config);
     }
     
-    @RequestMapping(value = "delete", method = RequestMethod.POST)
+    @POST
+    @Path("/delete")
+    @Consumes(MediaType.APPLICATION_JSON)
     public void delete(final RegistryCenterConfiguration config) {
         regCenterService.delete(config.getName());
     }
     
-    @RequestMapping(value = "connect", method = RequestMethod.POST)
-    public boolean connect(final RegistryCenterConfiguration config, final HttpSession session) {
-        return setRegistryCenterNameToSession(regCenterService.load(config.getName()), session);
+    @POST
+    @Path("/connect")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    public Boolean connect(final RegistryCenterConfiguration config, final @Context HttpServletRequest request) {
+        return setRegistryCenterNameToSession(regCenterService.load(config.getName()), request.getSession());
     }
     
     private boolean setRegistryCenterNameToSession(final RegistryCenterConfiguration regCenterConfig, final HttpSession session) {
         session.setAttribute(REG_CENTER_CONFIG_KEY, regCenterConfig);
         try {
             RegistryCenterFactory.createCoordinatorRegistryCenter(regCenterConfig.getZkAddressList(), regCenterConfig.getNamespace(), Optional.fromNullable(regCenterConfig.getDigest()));
+            SessionRegistryCenterConfiguration.setRegistryCenterConfiguration((RegistryCenterConfiguration) session.getAttribute(REG_CENTER_CONFIG_KEY));
         } catch (final RegException ex) {
             return false;
         }