+++ elastic-job-lite/elastic-job-lite-console/src/main/java/com/dangdang/ddframe/job/lite/console/service/impl/JobAPIServiceImpl.java
@@ -26,9 +26,7 @@ import com.dangdang.ddframe.job.lite.lifecycle.api.JobSettingsAPI;
 import com.dangdang.ddframe.job.lite.lifecycle.api.JobStatisticsAPI;
 import com.dangdang.ddframe.job.lite.lifecycle.api.ServerStatisticsAPI;
 import com.google.common.base.Optional;
-import org.springframework.stereotype.Service;
 
-@Service
 public final class JobAPIServiceImpl implements JobAPIService {
     
     @Override